+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.22.0] - 2025-05-15
+
+### Added
+
+- Support for multiple focal planes for files read with `OpenTile` source.
+
+### Fixed
+
+- Only converting the highest focal plane for files read with `OpenTile` source.
+
 ## [0.21.1] - 2025-05-14
 
 ### Fixed
@@ -351,8 +361,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of wsidicomizer
 
-[Unreleased]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.21.1..HEAD
-[0.21.1]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.20.1..0.21.1
+[Unreleased]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.22.0..HEAD
+[0.22.0]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.21.1..0.22.0
+[0.21.1]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.21.0..0.21.1
 [0.21.0]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.20.0..0.21.0
 [0.20.0]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.19.0..0.20.0
 [0.19.0]: https://github.com/imi-bigpicture/wsidicomizer/compare/0.18.0..0.19.0